Remove extra backendmessage table
It took me a *while* but I think I know what it is happening.

There is an import (`from .outgoing import DatabaseBackend`) in `rapidsms.backends.database.__init__.py`. When Django loads apps, it first tries to import the module. That triggers that import, which itself imports the `models.py` file, which means that models get loaded before the AppConfig is set up. Somehow this results in the models being attached to the `rapidsms` app_label instead of the `database` app_label.

I've been able to fix this by removing the `__init__.py` import, but I think that might break people's code. I've also been able to fix it by manually setting the `app_label` in the model's Meta class, which is what I've done in this PR.

The data in these tables is not critical, since it's used only for the Message Tester, or for other 'testing' purposes, but I've written a migration to migrate any data that ended up in the 'rapidsms' table so that it ends up in the 'database' table.

Addresses #460